+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Improved default metadata using `piletConfig` from *package.json* in `pilet debug` (#458)
 - Updated to React v18 (#501)
 - Updated `importmap` with `inherit` key
+- Updated `kras` to `v0.15`
 - Moved templates to dedicated repository (#458)
 - Changed the default bundler to be `piral-cli-webpack5` (#469)
 - Changed webpack configs to use `oneOf` for assets (#451)